legacy/requirements.txt
Flask==3.0.0
riotwatcher==3.3.1
pandas==2.1.3
scikit-learn==1.3.2
gunicorn==21.2.0
python-dotenv==1.0.0
numpy==1.26.2
